fix: added a common component for buttons
