+++ src/components/Home.tsx
@@ -1,21 +1,7 @@
-import Layout from "@/pages/layout";
-import { Button } from "@mui/material";
-import { motion, AnimatePresence } from "framer-motion";
+import { AnimatePresence, motion } from "framer-motion";
 import { useRouter } from "next/router";
 import { useEffect, useState } from "react";
-
-const buttonVariants = {
-  hover: {
-    scale: 1.1,
-    textShadow: "0px 0px 8px rgb(255,255,255)",
-    boxShadow: "0px 0px 8px rgb(255,255,255)",
-    transition: {
-      duration: 0.3,
-      repeat: Infinity,
-      repeatType: "reverse",
-    },
-  },
-};
+import StyledButtons from "./Buttons";
 
 const containerVariants = {
   hidden: {
@@ -39,7 +25,6 @@ const containerVariants = {
 };
 
 const HomeComponent = () => {
-  const MotionButton = motion(Button);
   const router = useRouter();
   const [isVisible, setIsVisible] = useState(true);
 
@@ -83,9 +68,9 @@ const HomeComponent = () => {
             Welcome to Pizza Joint
           </h2>
 
-          <MotionButton
-            variants={buttonVariants as any}
-            whileHover="hover"
+          <StyledButtons
+            name={"Create Your Pizza"}
+            handleClick={handleNavigation}
             style={{
               color: "white",
               background: "transparent",
@@ -95,9 +80,7 @@ const HomeComponent = () => {
               borderRadius: "50px",
               margin: "30px 20px",
             }}
-            onClick={handleNavigation}>
-            Create Your Pizza
-          </MotionButton>
+          />
         </motion.div>
       )}
     </AnimatePresence>
